pom.xml: Use GitHub for plugin's documentation on plugins.jenkins.io
The plugin description on the plugin page is short and incomplete. This
change will replace the plugin page with a better page based on the
readme.md file from GitHub. The change will take effect when the next
version of the plugin is released.

See https://groups.google.com/forum/#!topic/jenkinsci-dev/VSdfVMDIW-A and
https://issues.jenkins-ci.org/browse/WEBSITE-406 for details.